refactor: split frontend api into modules
